feat: ad types to board.service.js
